+++ src/wbs/clients/apn/chat/adult/daemon/ChatAdultDeliveryHandler.java
@@ -62,7 +62,7 @@ class ChatAdultDeliveryHandler
 	public
 	void handle (
 			int deliveryId,
-			Integer ref) {
+			Long ref) {
 
 		@Cleanup
 		Transaction transaction =
