+++ src/wbs/clients/apn/chat/core/daemon/ChatLogoffCommand.java
@@ -85,7 +85,7 @@ class ChatLogoffCommand
 	CommandRec command;
 
 	@Getter @Setter
-	Optional<Integer> commandRef;
+	Optional<Long> commandRef;
 
 	@Getter @Setter
 	String rest;
